+++ Sources/SecurityHandler/StoredDataModel.xcdatamodeld/StoredDataModel.xcdatamodel/contents
@@ -3,5 +3,6 @@
     <entity name="StoredDataModel" representedClassName="StoredDataModel" syncable="YES" codeGenerationType="class">
         <attribute name="data" optional="YES" attributeType="Binary" allowsCloudEncryption="YES"/>
         <attribute name="date" attributeType="Date" usesScalarValueType="NO"/>
+        <attribute name="expirationDuration" optional="YES" attributeType="Double" defaultValueString="0.0" usesScalarValueType="YES"/>
     </entity>
 </model>
